+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 * `new_node` no longer requires a code argument
 * `repr(database)` returns something useful
+* Remove fasteners and write locks
 
 ## 4.0.DEV18 (2022-08-19)